+++ src/main/java/http/HttpRequest.java
@@ -1,7 +1,8 @@
-package webserver;
+package http;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.RequestHandler;
 
 import java.io.*;
 import java.net.URLDecoder;
