+++ src/main/java/webserver/RequestHandler.java
@@ -7,6 +7,9 @@ import java.util.Map;
 
 import controller.Controller;
 import controller.UserController;
+import http.HttpRequest;
+import http.HttpResponse;
+import http.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
